src/home_assistant/requirements.txt
langchain-openai
crewai
langgraph
a2a-sdk